pyproject.toml
[tool.poetry]
name = "video-diet"
version = "0.1.0"
description = ""
authors = ["hian <hiancdtrsnm@gmail.com>"]
readme = "README.md"

[tool.poetry.scripts]
video-diet = "video_diet.main:app"

[tool.poetry.dependencies]
python = "^3.8"
typer = "^0.3.2"
ffmpeg-python = "^0.2.0"
filetype = "^1.0.7"
ffprobe-python = "^1.0.3"
enlighten = "^1.6.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
